add getSyncReadData with uint8_t

+++ dynamixel_workbench_controllers/include/dynamixel_workbench_controllers/dynamixel_workbench_controllers.h
@@ -102,7 +102,8 @@ class DynamixelController
   double pub_period_;
 
   bool is_moving_;
-  int32_t get_temperature[];
+  // int32_t get_temperature[];
+  uint8_t get_temperature[];
 
  public:
   DynamixelController();
